requirements.txt
PySide6==6.9.1
PyPDF2==3.0.1
pdfminer.six==20250506